+++ src/App.tsx
@@ -1,7 +1,7 @@
 import PrismComponents from "../lib/components.json";
 import { useLayoutEffect, useRef } from "react";
-import { theme, ui } from "./theme";
 import { usePersistentState } from "./usePersistentState";
+import { Theme, themes } from "./themes";
 
 const initialLang = "tsx";
 const initialCode = `\
@@ -52,13 +52,9 @@ const languages = sortObject(
     })
   )
 );
-
-const themeNames = sortObject({
-  ocean: "Ocean",
-  basic: "Basic",
-  nature: "Nature",
-  miasma: "Miasma",
-} as const);
+// "HTML" is called "Markup" by default which is not very descriptive
+languages.markup = "HTML";
+console.log(languages);
 
 function* walk(root: HTMLElement): Generator<HTMLElement> {
   yield root;
@@ -72,20 +68,33 @@ function* walk(root: HTMLElement): Generator<HTMLElement> {
   }
 }
 
-type ThemeName = keyof typeof themeNames;
+type ThemeName = keyof typeof themes;
+
+function isThemeID(x: unknown): x is ThemeName {
+  return typeof x === "string" && themes.hasOwnProperty(x);
+}
 
 export function App(): JSX.Element {
   const preRef = useRef<HTMLPreElement>(null);
   const proseRef = useRef<HTMLDivElement>(null);
   const [lang, setLang] = usePersistentState("language", initialLang);
   const [code, setCode] = usePersistentState("code", initialCode);
+  const [themeID, setThemeID] = usePersistentState(
+    "v2.theme",
+    themes.toxic._meta.id
+  );
   const [includeLink, setIncludeLink] = usePersistentState("includeLink", true);
+
+  const themeKey = isThemeID(themeID) ? themeID : "toxic";
+  const theme: Theme = themes[themeKey] || themes.toxic;
 
   function inlineStyles(root: HTMLElement) {
     root.classList.add("_root");
     for (const elem of walk(root)) {
       for (const cls of elem.classList) {
         const themeObj = theme[cls] || {};
+        // Reset styles to avoid leaking styles between themes
+        elem.setAttribute("style", "");
         for (const [key, val] of Object.entries(themeObj)) {
           elem.style.setProperty(key, val);
         }
@@ -117,7 +126,9 @@ export function App(): JSX.Element {
     input.type = "file";
     input.onchange = async () => {
       const [file] = input.files || [];
-      if (!file) return;
+      if (!file) {
+        return;
+      }
       const text = await file.text();
       setCode(text);
     };
@@ -175,22 +186,47 @@ export function App(): JSX.Element {
           }}
         />
       </div>
-      <div className="flex flex-wrap gap3 items-center">
-        <button className="bit-button" type="button" onClick={copyAsHTML}>
-          Copy as HTML
-        </button>
-        <label className="bit-field flex flex-wrap gap1 items-center">
-          <input
-            className="bit-checkbox"
-            type="checkbox"
-            checked={includeLink}
+      <div className="flex flex-wrap gap3 items-end">
+        <div className="flex flex-column gap1">
+          <label className="bit-label" htmlFor="form-theme">
+            Theme
+          </label>
+          <select
+            id="form-theme"
+            className="bit-select"
+            value={themeID}
             onChange={(event) => {
-              setIncludeLink(event.target.checked);
+              setThemeID(event.target.value);
             }}
-          />
-          <div>Include link to codehost</div>
-        </label>
+          >
+            {Object.entries(themes).map(([key, val]) => {
+              return (
+                <option key={key} value={key}>
+                  {val._meta.name}
+                </option>
+              );
+            })}
+          </select>
+        </div>
+        <div className="flex-auto" />
+        <div className="flex flex-wrap gap3 items-center">
+          <label className="bit-field flex flex-wrap gap1 items-center">
+            <input
+              className="bit-checkbox"
+              type="checkbox"
+              checked={includeLink}
+              onChange={(event) => {
+                setIncludeLink(event.target.checked);
+              }}
+            />
+            <div>Include link to codehost</div>
+          </label>
+          <button className="bit-button" type="button" onClick={copyAsHTML}>
+            Copy as HTML
+          </button>
+        </div>
       </div>
+
       <output className="code-output">
         <article>
           <header>
@@ -202,18 +238,7 @@ export function App(): JSX.Element {
             <div className="prose" ref={proseRef}>
               <pre className="_root" ref={preRef} />
               {includeLink && (
-                <div
-                  style={{
-                    margin: "-0.75rem",
-                    marginTop: "-1.25rem",
-                    marginBottom: "1.25rem",
-                    padding: "0.25rem 0.5rem",
-                    fontSize: "smaller",
-                    textAlign: "right",
-                    background: ui.bg1,
-                    color: ui.fg,
-                  }}
-                >
+                <div style={theme._footer}>
                   syntax highlighting by{" "}
                   <a
                     href="https://codehost.wavebeem.com"